# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "twp server api.",
        "title": "twp API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.o"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/coupon": {
            "get": {
                "description": "Get all global coupons .",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Coupon"
                ],
                "summary": "Admin Get Coupon",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Begin index",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.GetGlobalCouponsRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Add global coupon.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Coupon"
                ],
                "summary": "Admin Add Coupon",
                "parameters": [
                    {
                        "description": "Coupon",
                        "name": "coupon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.AddCouponParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.AddCouponRow"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/admin/coupon/{id}": {
            "get": {
                "description": "Get coupon details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Coupon",
                    "Shop"
                ],
                "summary": "Admin Get Coupon Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Coupon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.GetGlobalCouponDetailRow"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete coupon (include shops').",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Coupon"
                ],
                "summary": "Admin Delete Coupon",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Coupon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Edit global coupon. All the coupon properties are required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Coupon"
                ],
                "summary": "Admin Edit Coupon",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Coupon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Coupon",
                        "name": "coupon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.PrettierCoupon"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.EditCouponRow"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/admin/report": {
            "get": {
                "description": "Get site report (top 3 sellers and total amount).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Report"
                ],
                "summary": "Admin Get Site Report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start year/month",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.adminReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.PrettierCoupon": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/db.CouponType"
                }
            }
        },
        "admin.adminReport": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "integer"
                },
                "sellers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.GetTopSellerRow"
                    }
                },
                "total": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "db.AddCouponParams": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "expire_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/db.CouponScope"
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/db.CouponType"
                }
            }
        },
        "db.AddCouponRow": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "expire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/db.CouponScope"
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/db.CouponType"
                }
            }
        },
        "db.CouponScope": {
            "type": "string",
            "enum": [
                "global",
                "shop"
            ],
            "x-enum-varnames": [
                "CouponScopeGlobal",
                "CouponScopeShop"
            ]
        },
...
        "db.CouponType": {
            "type": "string",
            "enum": [
                "percentage",
                "fixed",
                "shipping"
            ],
            "x-enum-varnames": [
                "CouponTypePercentage",
                "CouponTypeFixed",
                "CouponTypeShipping"
            ]
        },
        "db.EditCouponRow": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "expire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/db.CouponScope"
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/db.CouponType"
                }
            }
        },
...
        "db.GetGlobalCouponDetailRow": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "expire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/db.CouponScope"
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/db.CouponType"
                }
            }
        },
        "db.GetGlobalCouponsRow": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "expire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/db.CouponScope"
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/db.CouponType"
                }
            }
        },
...
        "db.GetTopSellerRow": {
            "type": "object",
            "properties": {
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "seller_name": {
                    "type": "string"
                },
                "total_sales": {
                    "type": "integer"
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}